requirements.txt
langchain
langchain-experimental
langchain-openai
openai
python-environ
psycopg2
